BugFix: ADProjectileBase SetDamage의 인자가 더 작은 범위인 것을 수정

+++ Source/AbyssDiverUnderWorld/Projectile/ADProjectileBase.h
@@ -71,7 +71,7 @@ private:
 #pragma region Getter, Setter
 public:
 	UProjectileMovementComponent* GetProjectileMovementComp() const { return ProjectileMovementComp; }
-	void SetDamage(int8 DamageAmount) { Damage = DamageAmount; };
+	void SetDamage(int16 DamageAmount) { Damage = DamageAmount; };
 
 #pragma endregion
 };
